+++ lib/dom.ts
